omnisuite-dashboard/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
requests